requirements-local.txt
# VidSumGNN Requirements for Local RTX 3070 Setup
# Install with: pip install -r requirements-local.txt

# PyTorch with CUDA 12.1 support (RTX 3070)
# Install separately first with:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# PyTorch Geometric (install after PyTorch)
# pip install torch-geometric torch-scatter torch-sparse -f https://data.pyg.org/whl/torch-2.1.0+cu121.html
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.17

# Transformers and ML
transformers>=4.35.0
accelerate>=0.24.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0

# Video Processing
scenedetect[opencv]>=0.6.0
ffmpeg-python>=0.2.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0

# Jupyter (optional, for notebook)
jupyterlab>=4.0.0
ipywidgets>=8.1.0
notebook>=7.0.0

# Development (optional)
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
